Code Restructuring Proposal
So, since we are considering adding data structures for managing linear
algebra, we might want to consider re-iterating out ideas on how to
organize the code to better handle a lot of additions to the code. I
prefer a structure roughly similar to how d3.js structures there code.
Hence, the proposed commit (which also has some numpy influence for the
first iteration). Have an idea? Refactor to roughly your idea and then
we can compare before there's a scaling issue.